+++ code/src/skiplist.cc
@@ -1,10 +1,12 @@
+#include "skiplist.h"
+
 namespace bilsmtree {
 
 SkipList::SkipList() {
   head_ = new ListNode();
   head_->level_ = 0;
   head_->forward_ = new (ListNode*)[MAXLEVEL];
-  for (int i = 0; i < MAXLEVEL; ++ i)
+  for (size_t i = 0; i < MAXLEVEL; ++ i)
     head_->forward_[i] = NULL;
   data_size_ = 0;
 }
@@ -19,7 +21,7 @@ bool SkipList::IsFull() {
 
 Slice SkipList::Find(const Slice& key) {
   ListNode* p = head_;
-  for (int i = head_->level_ - 1; i >= 0; -- i) {
+  for (size_t i = head_->level_ - 1; i >= 0; -- i) {
     while (p->forward_[i].kv_.key_.compare(key) < 0)
       p = p->forward_[i];
   }
@@ -32,7 +34,7 @@ Slice SkipList::Find(const Slice& key) {
 void SkipList::Insert(const KV& kv) {
   ListNode* p = head_;
   ListNode* update_[MAXLEVEL];
-  for (int i = head_->level_ - 1; i >= 0; -- i) {
+  for (size_t i = head_->level_ - 1; i >= 0; -- i) {
     while (p->forward_[i].kv_.key_.compare(kv.key_) < 0)
       p = p->forward_[i];
     update_[i] = p;
@@ -47,11 +49,11 @@ void SkipList::Insert(const KV& kv) {
     q->level_ = GenerateLevel();
     q->forward_ = new (ListNode*)[q->level_];
     if (q->level_ > head_->level_) {
-      for (int i = head_->level_; i < q->level_; ++ i)
+      for (size_t i = head_->level_; i < q->level_; ++ i)
         update_[i] = head_;
       head_->level_ = q->level_;
     }
-    for (int i = 0; i < q->level_; ++ i) {
+    for (size_t i = 0; i < q->level_; ++ i) {
       q->forward_ = update_[i]->forward_[i];
       update_[i]->forward_[i] = q;
     }
@@ -61,14 +63,14 @@ void SkipList::Insert(const KV& kv) {
 void SkipList::Delete(const Slice& key) {
   ListNode* p = head_;
   ListNode* update_[MAXLEVEL];
-  for (int i = head_->level_ - 1; i >= 0; -- i) {
+  for (size_t i = head_->level_ - 1; i >= 0; -- i) {
     while (p->forward_[i].kv_.key_.compare(key) < 0)
       p = p -> forward_[i];
     update_[i] = p;
   }
   p = p->forward_[0];
   if (p->kv_.key_.compare(key) == 0) {
-    for (int i = 0; i < head_->level_; ++ i) {
+    for (size_t i = 0; i < head_->level_; ++ i) {
       if (update_[i]->forward_[i] != p)
         break;
       update_[i]->forward_[i] = p->forward_[i];
@@ -79,16 +81,15 @@ void SkipList::Delete(const Slice& key) {
   }
 }
 
-// TODO: NEED TO REDESIGN
 std::vector<KV> SkipList::GetAll() {
   std::vector<KV> res_;
-  for (int i = 0; i < data_size_; ++ i)
+  for (size_t i = 0; i < data_size_; ++ i)
     res_.push_back(head_[i].kv_);
   return res_;
 }
 
-int SkipList::GenerateLevel() {
-  int level_ = 1;
+size_t SkipList::GenerateLevel() {
+  size_t level_ = 1;
   while ((rand() * 1.0 / INT_MAX) < PROB && level_ < MAXLEVEL) 
     level_ = level_ + 1;
   return level_;
